PROJECT_NAME variable is now set immediately
In common.mk, the PROJECT_NAME variable was being set with the "="
operator. This is known as a recursively expanded variable: its value
is determined at the time it is used and its value can change. In the
case of PROJECT_NAME, since the value is based off the current
directory, it's unsafe to use a recursively expanded variable since
someone could change the current directory. Instead, I now use a simply
expanded variable: its value is determined at declaration time.